+++ core/cache.ts
@@ -4,74 +4,74 @@ import { md5 } from "./utils/digest.ts";
 
 export interface Options {
   /** The folder to load the files from */
-  folder?: string;
+  folder: string;
 }
 
 /**
- * Class to cache the content transformations (like transform_images manipulations)
+ * Class to manage cache in the _cache folder
  */
 export default class Cache {
   #folder: string;
 
-  constructor(options: Options = {}) {
-    this.#folder = options.folder || "_cache";
+  constructor(options: Options) {
+    this.#folder = options.folder;
   }
 
+  /** Save some content in the cache folder */
   async set(
-    content: string | Uint8Array,
-    key: unknown,
+    key: unknown[],
     result: string | Uint8Array,
   ): Promise<void> {
-    const [dir, file] = await paths(content, key);
+    const path = await this.getPath(key);
 
-    await ensureDir(posix.join(this.#folder, dir));
+    await ensureDir(posix.dirname(path));
 
     if (typeof result === "string") {
-      await Deno.writeTextFile(posix.join(this.#folder, dir, file), result);
+      Deno.writeTextFileSync(path, result);
     } else {
-      await Deno.writeFile(posix.join(this.#folder, dir, file), result);
+      Deno.writeFileSync(path, result);
     }
   }
 
-  async getFile(
-    content: Uint8Array | string,
-    key: unknown,
-  ): Promise<string> {
-    const [dir, file] = await paths(content, key);
-
-    return posix.join(this.#folder, dir, file);
-  }
-
-  async get(
-    content: Uint8Array | string,
-    key: unknown,
-  ): Promise<Uint8Array | undefined> {
+  /** Remove content from the cache folder */
+  async remove(key: unknown[]): Promise<void> {
     try {
-      return await Deno.readFile(await this.getFile(content, key));
+      Deno.removeSync(await this.getPath(key));
     } catch {
       // Ignore
     }
   }
 
-  async getText(content: string, key: unknown): Promise<string | undefined> {
+  async getPath(key: unknown[]): Promise<string> {
+    const paths = await Promise.all(key.map((value) => {
+      if (value instanceof Uint8Array || typeof value === "string") {
+        return md5(value);
+      }
+      return md5(JSON.stringify(value));
+    }));
+    return posix.join(this.#folder, ...paths);
+  }
+
+  /** Get the content from the cache folder as Uint8Array */
+  async getBytes(key: unknown[]): Promise<Uint8Array | undefined> {
     try {
-      return await Deno.readTextFile(await this.getFile(content, key));
+      return Deno.readFileSync(await this.getPath(key));
     } catch {
       // Ignore
     }
   }
 
+  /** Get the content from the cache folder as string */
+  async getText(key: unknown[]): Promise<string | undefined> {
+    try {
+      return Deno.readTextFileSync(await this.getPath(key));
+    } catch {
+      // Ignore
+    }
+  }
+
+  /** Empty the cache folder */
   async clear(): Promise<void> {
     await emptyDir(this.#folder);
   }
-}
-
-function paths(
-  content: string | Uint8Array,
-  key: unknown,
-): Promise<[string, string]> {
-  return Promise.all([
-    md5(content),
-    md5(JSON.stringify(key)),
-  ]);
 }